feat: 회원 정보 페이지 재구조화 및 기능 개발
