+++ package.json
@@ -35,6 +35,7 @@
     "recharts": "^3.0.2",
     "remark-gfm": "^4.0.1",
     "sass": "^1.83.0",
+    "tailwind-merge": "^3.4.0",
     "valtio": "^2.1.5",
     "web-vitals": "^2.1.4"
   },
