+++ src/pages/myPage/AccountInfo/AccountDelete/AccountDelete.scss
@@ -1,4 +1,4 @@
-@use "../../../styles/_responsive.scss" as r;
+@use "@/styles/_responsive.scss" as r;
 
 .account-delete {
   width: 414px;
