+++ src/pages/myPage/AccountInfo/AccountInfoChange.jsx
@@ -1,26 +1,27 @@
 import { useContext, useState } from "react";
 
 import AccountInfoChangeEnter from "./AccountInfoChange/ChangeEnter.jsx";
-import AccountInfoChangeMain from "./AccountInfoChange/ChangeMain.jsx";
+import AccountInfoChangePassword from "./AccountInfoChange/ChangePassword.jsx";
 import AccountDelete from "./AccountDelete/AccountDelete.jsx";
 import AccountDeleteSuccess from "./AccountDelete/AccountDeleteSuccess.jsx";
-import MyPageMenu from "../../components/myPage/MyPageMenu.jsx";
+import MyPageMenu from "@/components/myPage/MyPageMenu.jsx";
 
-import AuthContext from "../../contexts/AuthContext.jsx";
+import AuthContext from "@/contexts/AuthContext.jsx";
 
-import "./MyPageContentsDefault.scss";
+import "../MyPageContentsDefault.scss";
 import "./AccountInfoChange.css";
-import "./../../styles/utilities.css";
+import "@/styles/utilities.css";
 
+/**
+ * @deprecated AccountInfo.tsx 사용
+ */
 const AccountInfoChange = () => {
   // 회원 정보 수정 진입
   const [changeShowPermission, setChangeShowPermission] = useState(false);
 
   // 계정 삭제
-  const [isDeleteAccountBtnClicked, setIsDeleteAccountBtnClicked] =
-    useState(false);
-  const [isAccountSuccessfullyDeleted, setIsAccountSuccessfullyDeleted] =
-    useState(false);
+  const [isDeleteAccountBtnClicked, setIsDeleteAccountBtnClicked] = useState(false);
+  const [isAccountSuccessfullyDeleted, setIsAccountSuccessfullyDeleted] = useState(false);
 
   const { userNickname } = useContext(AuthContext);
 
@@ -32,22 +33,16 @@ const AccountInfoChange = () => {
           <div className="content-side-inside">
             {!changeShowPermission ? (
               /* 회원 정보 수정 진입 페이지 */
-              <AccountInfoChangeEnter
-                setChangeShowPermission={setChangeShowPermission}
-              />
+              <AccountInfoChangeEnter setChangeShowPermission={setChangeShowPermission} />
             ) : isAccountSuccessfullyDeleted ? (
               /* 계정 삭제 완료 페이지 */
               <AccountDeleteSuccess />
             ) : isDeleteAccountBtnClicked ? (
               /* 계정 삭제 페이지 */
-              <AccountDelete
-                setIsAccountSuccessfullyDeleted={
-                  setIsAccountSuccessfullyDeleted
-                }
-              />
+              <AccountDelete setIsAccountSuccessfullyDeleted={setIsAccountSuccessfullyDeleted} />
             ) : (
               // 회원 정보 수정 페이지
-              <AccountInfoChangeMain
+              <AccountInfoChangePassword
                 setIsDeleteAccountBtnClicked={setIsDeleteAccountBtnClicked}
               />
             )}
