+++ src/pages/myPage/AccountInfo/AccountInfoChange/ChangeEnter.jsx
@@ -1,14 +1,18 @@
 import { api } from "@/api/api";
-import { useState } from "react";
+import React, { useState } from "react";
 
-import { AuthPwInputField } from "../../../components/inputField";
-import EnterForm from "../../../components/EnterForm";
+import { AuthPwInputField } from "@/components/inputField";
+import EnterForm from "@/components/EnterForm";
 
 import useWindowDimensions from "@/hooks/useWindowDimensions";
 
 import "./ChangeEnter.scss";
-import "./../../../styles/text.css";
+import "@/styles/text.css";
 
+/**
+ * 비밀번호 변경 페이지 진입 페이지
+ * @param {Dispatch<SetStateAction<boolean>>} setChangeShowPermission - 비밀번호 변경 페이지 진입 페이지 설정 함수
+ */
 const AccountInfoChangeEnter = ({ setChangeShowPermission }) => {
   const [typedPassword, setTypedPassword] = useState("");
   const [pwValid, setPwValid] = useState(false);
@@ -41,16 +45,11 @@ const AccountInfoChangeEnter = ({ setChangeShowPermission }) => {
     <EnterForm onSubmit={onClickInputBtn}>
       <div className="info-change-enter">
         {/* 진입 페이지 */}
-        <h4 className={`${isSmallMobile ? "p-medium-bold" : "h4-bold"}`}>
-          회원 정보 수정
-        </h4>
+        <h4 className={`${isSmallMobile ? "p-medium-bold" : "h4-bold"}`}>회원 정보 수정</h4>
         <h6 className={`${isSmallMobile ? "p-xs-medium" : "p-medium-medium"}`}>
           회원 정보 수정을 위해서 비밀번호를 다시 한 번 입력해주세요.
         </h6>
-        <div
-          className={` a-items-start ${isSmallMobile ? "w-[280px]" : ""}`}
-          id="enter-input"
-        >
+        <div className={` a-items-start ${isSmallMobile ? "w-[280px]" : ""}`} id="enter-input">
           <AuthPwInputField
             placeholder="비밀번호를 입력해주세요."
             value={typedPassword}
