+++ src/pages/myPage/AccountInfo/AccountInfoChange/ChangePassword.scss
@@ -1,7 +1,7 @@
-@use "../../../styles/responsive" as r;
+@use "@/styles/responsive" as r;
 .info-change-main {
   width: 414px;
-  margin: 0 auto;
+  // margin: 0 auto;
   @include r.media-small-mobile {
     width: 280px;
   }
